complete cart and checkout

+++ store/views.py
@@ -1,10 +1,12 @@
+from itertools import product
 from django.shortcuts import render
 from django.http import JsonResponse, HttpResponseNotFound
 import json, datetime
 from django.views import View
 from django.views.generic import ListView, CreateView
 
-from store.models import Order, OrderItem, Product, ShippingAddress
+from store.models import *
+from store.utils import cartData, cookieCart, guestOrder
 
 class store(ListView):
 
@@ -15,12 +17,7 @@ class store(ListView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
 
-        if self.request.user.is_authenticated:
-            customer = self.request.user.customer
-            order, created = Order.objects.get_or_create(customer=customer, complete=False)
-            context['order'] = order
-        else:
-            context['order'] = {'get_cart_items':0, 'shipping':False}
+        context['order'] = cartData(self.request, False)['order']
         return context
 
 
@@ -33,58 +30,72 @@ class updateItem(View):
         productId = data['productId']
         action = data['action']
 
-        customer = request.user.customer
-        product = Product.objects.get(pk=productId)
-        order, created = Order.objects.get_or_create(customer=customer, complete=False)
+        # authenticated user
+        if request.user.is_authenticated:
+            customer = request.user.customer
+            order, created = Order.objects.get_or_create(customer=customer, complete=False)
 
-        orderItem, created = OrderItem.objects.get_or_create(order=order, product=product)
-        if action == 'add':
-            orderItem.quantity += 1
-        elif action == 'remove':
-            orderItem.quantity -= 1
-        orderItem.save()
+            product = Product.objects.get(pk=productId)
+            orderItem, created = OrderItem.objects.get_or_create(order=order, product=product)
+            if action == 'add':
+                orderItem.quantity += 1
+            elif action == 'remove':
+                orderItem.quantity -= 1
+            orderItem.save()
 
-        data = dict()
-        if orderItem.quantity <= 0:
-            orderItem.delete()
-            data['remove'] = True
-        else:          
+            if orderItem.quantity <= 0:
+                orderItem.delete()
+                data['remove'] = True
+            else:
+                data['total-price'] = order.get_cart_total
+                data['item-quantity'] = orderItem.quantity
+                data['item-price'] = orderItem.get_total
             data['total-items'] = order.get_cart_items
-            data['total-price'] = order.get_cart_total
-            data['item-id'] = orderItem.id
-            data['item-quantity'] = orderItem.quantity
-            data['item-price'] = orderItem.get_total
+        else:          
+            result = cookieCart(request.COOKIES['cart'])
+            
+            data['total-items'] = result['order']['get_cart_items']
+            data['total-price'] = result['order']['get_cart_total']
+
+            for item in result['items']:
+                if (item['product'].id == int(productId)):
+                    data['item-quantity'] = item['quantity']
+                    data['item-price'] = item['get_total']
+                    break
 
         return JsonResponse(data)
 
 
 def processOrder(request):
 
-    transactionId = datetime.datetime.now().timestamp()
     data = json.loads(request.body)
 
     if request.user.is_authenticated:
         # check the order data
         customer = request.user.customer
         order, created = Order.objects.get_or_create(customer=customer, complete=False)
-        order.transaction_id = str(transactionId)
-        total = order.get_cart_total
-        order.complete = True
-        order.save()
-
-        # check the shipping data
-        if order.shipping:
-            ShippingAddress.objects.create(
-                customer=customer,
-                order=order,
-                address=data['shipping']['address'],
-                city=data['shipping']['city'],
-                state=data['shipping']['state'],
-                zipcode=data['shipping']['zipcode'],
-            )
+        
     else:
         print('User is not logged in..')
+        customer, order = guestOrder(request.COOKIES['cart'], data['form'])
 
+    transactionId = datetime.datetime.now().timestamp()
+    order.transaction_id = str(transactionId)
+    # total = order.get_cart_total
+    order.complete = True
+    order.save()
+
+    # check the shipping data
+    if order.shipping:
+        ShippingAddress.objects.create(
+            customer=customer,
+            order=order,
+            address=data['shipping']['address'],
+            city=data['shipping']['city'],
+            state=data['shipping']['state'],
+            zipcode=data['shipping']['zipcode'],
+        )
+        
     return JsonResponse('payment submitted..', safe=False)
 
 # class processOrder(CreateView):
@@ -93,28 +104,11 @@ def processOrder(request):
 
 class cart(View):
     def get(self, request):
-
-        if request.user.is_authenticated:
-            customer = request.user.customer
-            order, created = Order.objects.get_or_create(customer=customer, complete=False)
-            items = order.orderitem_set.all()
-        else:
-            items = []
-            order = {'get_cart_items':0, 'get_cart_total':0, 'shipping':False}
-
-        context = {'items':items, 'order':order}
+        context = cartData(self.request)
         return render(request, 'store/cart.html', context)
 
 
 class checkout(View):
     def get(self, request):
-        if request.user.is_authenticated:
-            customer = request.user.customer
-            order, created = Order.objects.get_or_create(customer=customer, complete=False)
-            items = order.orderitem_set.all()
-        else:
-            items = []
-            order = {'get_cart_items':0, 'get_cart_total':0, 'shipping':False}
-
-        context = {'items':items, 'order':order}
+        context = cartData(self.request)
         return render(request, 'store/checkout.html', context)
